+++ activestorage/lib/active_storage/service/mirror_service.rb
@@ -49,6 +49,11 @@ module ActiveStorage
       perform_across_services :delete_prefixed, prefix
     end
 
+    def disk?
+      [ primary, *mirrors ].any? do |service|
+        service.try(:disk?)
+      end
+    end
 
     # Copy the file at the +key+ from the primary service to each of the mirrors where it doesn't already exist.
     def mirror(key, checksum:)